# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Nexus API
  version: 1.0.0
  description: |
    Project & Task Management API.
    All authenticated endpoints require a Bearer JWT access token.
    Access tokens expire in 15 minutes; use POST /auth/refresh to rotate.
servers:
  - url: https://api.nexus.io/api/v1
    description: Production
  - url: http://localhost:3001/api/v1
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:        { type: string, format: uuid }
        email:     { type: string, format: email }
        fullName:  { type: string }
        role:      { type: string, enum: [user, admin, superadmin] }
        avatarUrl: { type: string, nullable: true }
    TokenPair:
      type: object
      properties:
        accessToken:  { type: string }
        refreshToken: { type: string }
        expiresIn:    { type: integer, example: 900 }
        user:         { $ref: '#/components/schemas/User' }
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, fullName]
              properties:
                email:    { type: string, format: email }
                password: { type: string, minLength: 8 }
                fullName: { type: string, minLength: 2 }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/TokenPair' } } } }
        '409': { description: Email already registered }
        '422': { description: Validation error }
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Log in with email + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TokenPair' } } } }
        '401': { description: Invalid credentials }
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Rotate access + refresh token pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200': { description: New token pair }
        '401': { description: Invalid or reused refresh token }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Blacklist current access token
      responses:
        '200': { description: Logged out }
        '401': { description: Unauthenticated }
...
  /teams:
    get:
      tags: [Teams]
      summary: List my teams
      responses:
        '200': { description: Team list }
    post:
      tags: [Teams]
      summary: Create a team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '201': { description: Created }
  /teams/{teamId}/invite:
    post:
      tags: [Teams]
      summary: Invite a user by email (admin+)
      parameters:
        - { in: path, name: teamId, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
                role:  { type: string, enum: [viewer, member, admin] }
      responses:
        '200': { description: Member added }
        '404': { description: User not found }